# scheduler: Use heap-sort instead of quickselect