The HA scheduling algorithm introduced code that built a slice of all
nodes meeting the constraints, then selected the k best nodes from that
slice.

Building the slice can be expensive if there are a lot of nodes.

To improve performance, we can build a max-heap of nodes and pop all of
the elements to perform a heap sort. We never need to have more than k
items in the heap at any time, since the heap enables a trivial
optimization to check if a node is "better" than the root of the heap.
If it isn't, we can skip that node, and if it is, we can replace the
root of the heap with it.

This approach needs less memory allocation with k << len(nodes).

It also removes the dependency on quickselect.

Thanks to @LK4D4 for the idea.