@aaronlehmann 2016-09-19
Benchmark results:

```

benchmark                                           old ns/op       new ns/op       delta
BenchmarkScheduler1kNodes1kTasks-8                  99535349        98353639        -1.19%
BenchmarkScheduler1kNodes10kTasks-8                 546633387       548325249       +0.31%
BenchmarkScheduler1kNodes100kTasks-8                6060034233      6160128707      +1.65%
BenchmarkScheduler100kNodes1kTasks-8                1907288560      311341903       -83.68%
BenchmarkScheduler100kNodes100kTasks-8              6561291350      6670113753      +1.66%
BenchmarkScheduler100kNodes1MTasks-8                87268133282     81827858414     -6.23%
BenchmarkSchedulerConstraints1kNodes1kTasks-8       165552129       153443398       -7.31%
BenchmarkSchedulerConstraints1kNodes10kTasks-8      729888274       722834862       -0.97%
BenchmarkSchedulerConstraints1kNodes100kTasks-8     6605733833      6801978602      +2.97%
BenchmarkSchedulerConstraints5kNodes100kTasks-8     6662712088      6575577455      -1.31%

benchmark                                           old allocs     new allocs     delta
BenchmarkScheduler1kNodes1kTasks-8                  243095         244065         +0.40%
BenchmarkScheduler1kNodes10kTasks-8                 2760071        2760575        +0.02%
BenchmarkScheduler1kNodes100kTasks-8                31112694       31106768       -0.02%
BenchmarkScheduler100kNodes1kTasks-8                348289         349075         +0.23%
BenchmarkScheduler100kNodes100kTasks-8              31005741       31105897       +0.32%
BenchmarkScheduler100kNodes1MTasks-8                345009629      345098851      +0.03%
BenchmarkSchedulerConstraints1kNodes1kTasks-8       255069         257582         +0.99%
BenchmarkSchedulerConstraints1kNodes10kTasks-8      2879668        2880483        +0.03%
BenchmarkSchedulerConstraints1kNodes100kTasks-8     32308673       32314324       +0.02%
BenchmarkSchedulerConstraints5kNodes100kTasks-8     32164883       32168084       +0.01%

benchmark                                           old bytes       new bytes       delta
BenchmarkScheduler1kNodes1kTasks-8                  17889218        17944530        +0.31%
BenchmarkScheduler1kNodes10kTasks-8                 198313008       198338596       +0.01%
BenchmarkScheduler1kNodes100kTasks-8                2207604464      2206614656      -0.04%
BenchmarkScheduler100kNodes1kTasks-8                69665912        43189427        -38.00%
BenchmarkScheduler100kNodes100kTasks-8              2326743080      2332232632      +0.24%
BenchmarkScheduler100kNodes1MTasks-8                24970282280     24969463304     -0.00%
BenchmarkSchedulerConstraints1kNodes1kTasks-8       18517088        18688487        +0.93%
BenchmarkSchedulerConstraints1kNodes10kTasks-8      204626968       204619544       -0.00%
BenchmarkSchedulerConstraints1kNodes100kTasks-8     2269526128      2270519280      +0.04%
BenchmarkSchedulerConstraints5kNodes100kTasks-8     2275526616      2274922968      -0.03%
```

As expected, this is a big improvement if there are many more nodes than tasks being scheduled. Otherwise, the difference is minor.